pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genegan-pytorch-mps"
version = "0.1.0"
description = "GeneGAN PyTorch reproduction with Apple Silicon MPS support"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.23",
  "Pillow>=10.0",
  "opencv-python>=4.8",
  "pyyaml>=6.0",
  "tensorboard>=2.12",
  "torch>=2.1",
  "torchvision>=0.16",
  "tqdm>=4.66",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["genegan*"]
